Upgrade to kotlin 1.0.0-beta-1038

+++ src/main/kotlin/soy/frank/blog/models/Post.kt
@@ -8,7 +8,7 @@ import javax.persistence.*
 @Table(name = "posts")
 data class Post(
         @Id
-        val id: Long = -1,
+        val id: Int = -1,
         val title: String = "",
         val renderedBody: String = "",
         @Convert(converter = LocalDateTimePersistenceConverter::class)
